backend/requirements.txt
Django 
Pandas
numpy
pytest-django
sqlparse
djangorestframework
djangorestframework-simplejwt